+++ spec/setup_websockify_spec.rb
@@ -34,4 +34,15 @@ describe 'abiquo::setup_websockify' do
             :group => 'root'
         )
     end
+
+
+    it 'configures the websockify cron task' do
+      chef_run.converge('abiquo::install_websockify', described_recipe, 'abiquo::service')
+      expect(chef_run).to create_file("/etc/cron.d/novnc_tokens").with(
+          :content  => "* * * * * root /opt/websockify/novnc_tokens.rb -a #{chef_run.node['abiquo']['websockify']['api_url']} -u #{chef_run.node['abiquo']['websockify']['user']} -p #{chef_run.node['abiquo']['websockify']['pass']} -f /opt/websockify/config.vnc",
+          :owner    => 'root',
+          :group    => 'root',
+          :mode     => '0644'
+      )
+    end
 end