+++ test/integration/frontend/serverspec/frontend_packages_spec.rb
@@ -14,19 +14,19 @@
 
 require "#{ENV['BUSSER_ROOT']}/../kitchen/data/serverspec_helper"
 
-describe 'UI packages' do
+describe 'Front-end packages' do
     it 'has the system packages installed' do
         expect(package('cronie')).to be_installed
     end
 
     it 'has the abiquo packages installed' do
-        %w{ui tutorials}.each do |pkg|
+        %w{ui tutorials websockify}.each do |pkg|
             expect(package("abiquo-#{pkg}")).to be_installed
         end
     end
 
     it 'does not have other abiquo installed' do
-        %w{websockify server remote-services monolithic nodecollector}.each do |pkg|
+        %w{server remote-services monolithic nodecollector}.each do |pkg|
             expect(package("abiquo-#{pkg}")).to_not be_installed
         end
     end