+++ CHANGELOG.md
@@ -3,6 +3,9 @@ abiquo CHANGELOG
 
 ## 0.8.0
 
+* Simplified the ui configuration by leveraging only the 'ui_config' attribute.
+* Added ext_services profile.
+* Added front-end profile.
 * Configured the monitoring resources to be idempotent.
 * Fix watchtower schema creation during install and upgrade.
 * Upgrade tests to Centos 6.8.